Properly fix region alignement Fixes Avasam#21 Updated some doc max fps will show `...` while working
