+++ res/design.ui
@@ -172,7 +172,7 @@
     <property name="geometry">
      <rect>
       <x>92</x>
-      <y>255</y>
+      <y>254</y>
       <width>20</width>
       <height>20</height>
      </rect>
@@ -274,7 +274,7 @@
     <property name="geometry">
      <rect>
       <x>65</x>
-      <y>255</y>
+      <y>254</y>
       <width>26</width>
       <height>20</height>
      </rect>
